Silence some compiler warnings
Replaced the variadic macrot THROW(fmt, ...) with a C++11-compilant one:
Because ## __VA_ARGS__ is a GNU extension, a warning was generated with
-Wpedantic. By using __VA_ARGS__ without ##, no warning is emitted when
the ARFF headers are included in a C++11 project. With earlier language
versions, the GNU extension may be used.

Also removed two superfluous commas in enum declarations.

The long long typedef in arff_util.h still generated a warning in C++98,
but there is not much to do (portably) about it...